fix: keep Composer recipe validation browser-safe
